Added posts.js and users.js
Modified users.js and created posts.js to split up routes into a more intuitive naming scheme.
